military-3d-automation/requirements.txt
opencv-python>=4.10.0.84
pillow>=10.4.0
trimesh>=4.4.9
numpy>=2.1.0
